mdpi: collection of optimization for MDPI device
Gallery3D: Fix the picture caching system. Generate 2 files for cache: screennail and hires.
Max. resolutions  must be set according to memory and GPU capabilities with config.xml

Change-Id: I7d51ba2be97dfdc0ef536609594bc19a66fc57ae

Lowered te thumbnailer's waiting time from 2000ms to 500ms to decrease latency on startup

Furher lowered the thumbnailer's waiting time.

Made the quality factor adaptive to the VMMaxHeapSize
The quality factor is calculated from the formula (int) VMMaxHeapSize(mB)/12. That way, standard, non modified devices get images at four times the resolution (2x bigger in width and 2x bigger in height). Devices that have lower memory than 24mB won't be affected by the respatch.

Change-Id: I704c2fc5b5394c94c44cf8008f0564a38e4767fc

Calculate respatch on startup from App.java
The respatch related code has now been moved into the class App which is initiated on application load. This way, the factor doesn't have to be calculated on each image load, but is calculated only once (on startup) leading to some performance improvements.
Since the factor is now a part of the App class, it can be accessed from any part of the code with App.RESPATCH_FACTOR (which is now used instead of the old code in Utils.java).
Also, a comment explaining the respatch in short has been added.

Change-Id: Ie21ec63324bfb5697835fac15c9e74d5b4715aa3

Respatch :: Changed the divisor to 9 and commented related values
9 seems as the most appropriate value for the respatch (works on default devices and allows for high res images on custom ones).
Also added comments for App.RESPATCH_FACTOR (with a guide on how to tune the respatch, including how to disable it) and for App.RESPATCH_DIVISOR.

Change-Id: Id04f1e73e97e8d8914a85af5d1cbb0bd7d43f9b1

CacheService :: Prevent division by zero
It seems that there are some situations in which thumbnailWidth will be 0. Since thumbnailWidth is used as a divisor in writeBitmapToCache, this can lead to a nasty Force Close.